assigments_1_to_6_projects/8_qrcode_encode_decoder/pyproject.toml
[project]
name = "8-qrcode-encode-decoder"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "opencv-python-headless>=4.11.0.86",
    "pillow>=11.1.0",
    "pyzbar>=0.1.9",
    "qrcode>=8.0",
    "streamlit>=1.44.0",
]
